Implement word gift claiming via server action
